Changed the game components for managers and fixed many errors
Changed the components of the game folder for Managers that will update the
shared data of the entities added and store its specific data (they wrap the
systems)

Fixed orientation errors (never use the glm *= operator with quaternions!)

Changes the FreeImage texture library for SOIL

Changed the loaders folder: now the classes used for loading will use
the Loader name and the ones used for reading from files will use the
Reader name

Changed the fine collision detection between spheres so now we can use the
GJK Algorithm with all the colliders

+++ src/window/InputData.h
@@ -18,20 +18,22 @@ namespace window {
 		/** An array with the mouse pressed keys */
 		bool mMouseButtons[MAX_MOUSE_BUTTONS];
 
-		/** The Mouse X coordinate */
+		/** The Mouse X coordinate relative to the top left corner of the
+		 * window*/
 		float mMouseX;
 
-		/** The Mouse Y coordinate */
+		/** The Mouse Y coordinate relative to the top left corner of the
+		 * window */
 		float mMouseY;
 
 		/** Creates a new InputData */
 		InputData() : mMouseX(0.0f), mMouseY(0.0f)
 		{
-			for (unsigned int i = 0; i < MAX_KEYS; ++i) {
+			for (size_t i = 0; i < MAX_KEYS; ++i) {
 				mKeys[i] = false;
 			}
 
-			for (unsigned int i = 0; i < MAX_MOUSE_BUTTONS; ++i) {
+			for (size_t i = 0; i < MAX_MOUSE_BUTTONS; ++i) {
 				mMouseButtons[i] = false;
 			}
 		};
